AgriConnect/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "djaodjin-pages>=0.8.6",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "qrcode>=8.2",
    "streamlit>=1.49.1",
]
